api/requirements.txt
Django==2.1.14
django-cors-headers
djangorestframework == 3.10.3
gunicorn==19.3.0